requirements.txt
Django==1.10.3
mysqlclient==1.3.9
dj-database-url
dj-static==0.0.6
grpcio
grpcio-tools
Whoosh
